Make EP3 valid even if we receive 0 bytes; make whole USB buffer volatile
First this change makes the entire usb_state structure volatile.  Why?
Because since interrupts have effect on every single primitive in there
(including the buffers) we need to ensure we don't optimize any of that
out.

Second this fix adjust our EP3_Out callback to re-enable the receiving
of data if we receive nothing.  Previously we didn't do this and it could
lead us to get stuck if we tried to receive something and nothing was
there.